Move View to presantation directory
