+++ webapp/src/presentation/views/EntityEdit/components/MenuBar/MenuBar.tsx
