+++ webapp/src/presentation/views/NewEntityEdit/components/MenuBar/MenuBar.tsx
