+++ webapp/src/presentation/views/NewEntityEdit/components/MenuBar/index.ts
