+++ webapp/src/presentation/views/NewEntityEdit/components/PropertyMenu/PropertyMenu.tsx
