+++ webapp/src/presentation/views/NewEntityEdit/components/PropertyMenu/index.ts
